+++ packages/core/src/managers/index.ts
@@ -1,3 +1,3 @@
 export * from "./ArrowManager";
 export * from "./DraggableManager";
-export * from "./GestureManager";
+export * from "./ClickEventManager";